refactor(app): separate system updates from input handling
This commit moves the lazy screen update logic out of the handler loop and into
App::process_system_updates. The main loop still preserves the existing input
dispatch behavior, but non-input updates such as initial mailing list loading,
latest patchset loading, and empty bookmark fallback now live behind an App
boundary.

It also adds App::input_context as the state projection that will feed the
input mapper once the remaining handlers consume semantic InputEvent values.

This commit is part of the architecture's refactoring phase 8.
